Workaround for host resolution from cluster network
Pods are unable to resolve the master/worker nodes addresses since we're
using hosts. This breaks for example `metrics-server` as it's trying to
screpe the kubelets using the hostnames of the nodes.

Added the `/etc/hosts` file as a volume mount to the coredns deployment
and serve it using the `hosts` coredns plugin.